ci: bundle without pino plugin
