+++ packages/salesforcedx-vscode-apex-debugger/package.json
@@ -35,7 +35,6 @@
     "@types/node": "^20.0.0",
     "cross-env": "5.2.0",
     "esbuild": "0.25.0",
-    "esbuild-plugin-pino": "^2.2.2",
     "salesforcedx-vscode-apex": "64.10.1"
   },
   "extensionDependencies": [
@@ -45,7 +44,7 @@
   "scripts": {
     "bundle:extension": "npm run bundle:extension:build && npm run bundle:extension:copy",
     "bundle:extension:copy": "cp ../salesforcedx-apex-debugger/dist/apexDebug.js ./dist/",
-    "bundle:extension:build": "node ./esbuild.config.js",
+    "bundle:extension:build": "node ./esbuild.config.mjs",
     "vscode:prepublish": "npm prune --production",
     "vscode:package": "ts-node  ../../scripts/vsce-bundled-extension.ts",
     "vscode:sha256": "node ../../scripts/generate-sha256.js >> ../../SHA256",
