+++ packages/salesforcedx-vscode-apex-replay-debugger/package.json
@@ -37,7 +37,6 @@
     "@types/node": "^20.0.0",
     "cross-env": "5.2.0",
     "esbuild": "0.25.0",
-    "esbuild-plugin-pino": "^2.2.2",
     "salesforcedx-vscode-apex": "64.10.1",
     "salesforcedx-vscode-core": "64.10.1",
     "vscode-extension-telemetry": "^0.0.17"
@@ -49,7 +48,7 @@
   "scripts": {
     "bundle:extension": "npm run bundle:extension:build && npm run bundle:extension:copy",
     "bundle:extension:copy": "cp ../salesforcedx-apex-replay-debugger/dist/apexreplaydebug.js ./dist/",
-    "bundle:extension:build": "node ./esbuild.config.js",
+    "bundle:extension:build": "node ./esbuild.config.mjs",
     "vscode:prepublish": "npm prune --production",
     "vscode:package": "ts-node  ../../scripts/vsce-bundled-extension.ts",
     "vscode:sha256": "node ../../scripts/generate-sha256.js >> ../../SHA256",
