+++ packages/salesforcedx-vscode-apex/package.json
@@ -51,7 +51,6 @@
     "@types/node": "^20.0.0",
     "cross-env": "5.2.0",
     "esbuild": "0.25.0",
-    "esbuild-plugin-pino": "^2.2.2",
     "openapi-types": "12.1.3",
     "salesforcedx-vscode-core": "64.10.1"
   },
@@ -61,7 +60,7 @@
   "scripts": {
     "bundle:extension": "npm run bundle:extension:build && npm run bundle:extension:copy",
     "bundle:extension:copy": "cp ./out/apex-jorje-lsp.jar ./dist/",
-    "bundle:extension:build": "node ./esbuild.config.js",
+    "bundle:extension:build": "node ./esbuild.config.mjs",
     "vscode:package": "ts-node  ../../scripts/vsce-bundled-extension.ts",
     "vscode:sha256": "node ../../scripts/generate-sha256.js >> ../../SHA256",
     "vscode:publish": "node ../../scripts/publish-vsix.js",
